// lib/statistics/interface.ts
export interface IStatistics {
    year: {
        permissions: number;
        vacations: number;
        disabilities: number;
        overtimes: number;
        penalties: number;
        excusedAbsences: number;
        unexcusedAbsences: number;
    }
    period: {
        permissions: number;
        vacations: number;
        disabilities: number;
        overtimes: number;
        penalties: number;
        excusedAbsences: number;
        unexcusedAbsences: number;
    }
}